requirements.txt
psutil>=5.9.0
requests>=2.28.0
pyyaml>=6.0
schedule>=1.2.0
websocket-client>=1.4.0
python-systemd>=234; sys_platform == "linux"
setproctitle>=1.3.0
